refactor: rework folder organisation
